python/requirements.txt
PyGithub
argparse
pylint
